pyproject.toml
[project]
name = "multi-document-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.2",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.2",
    "uvicorn[standard]>=0.40.0",
]
